+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.10.12
+- Add the ability to create a shared memory with no value
+- And load data from a readable source into that shared memory
+
 # 1.10.11
 - add support type bytes or bytearray for share memory in BaseTask